==> analysis/04_land_use_dependent_dispersal_scripts/00_movebank_login.R <==
# Load packages ----------------------------------------------------------------
source(list.files("./R", full.names = T))

ipak("tidyverse")


# Movebank ---------------------
ipak("move")

curl_login <- movebankLogin(username = "FILLIN",
                            password = "FILLIN")
